[sil] Ban reborrows/guaranteed phis in Raw SIL.
We need reborrows to exist in SIL so we can mem2reg borrowed things in the
performance pipeline. Today, earlier in the pipeline, we never emit reborrows,
so in this commit I future proof this condition by banning reborrows in Raw SIL
via the SILVerifier. This simplifies the world that diagnostic passes need to
consider. If we need to re-enable support for reborrows in the diagnostic
pipeline, we can ban it later in the diagnostic pipeline.

NOTE: The test modifications involved me just explicitly marking these tests as
being in canonical SIL instead of Raw SIL so we don't trigger the verifier (Raw
SIL seems to be assumed if one doesn't specify the stage).