Pass swift and clang paths to build-script-impl
This patch updates the build-script-impl invocation to pass the paths to
the swift-tools and clang-tools from the toolchain to build-script-impl
if the tool is present. build-script-impl expects this to be to the path
containing the binary directly, so pulling the dirname from the absolute
path to the binary should be sufficient.