[ClangImporter] Use clang getDefaultLanguageStandard
This moves the logic for selecting the default language standard into
Clang, instead of duplicating it. This is intended to make the reliance
on CLANG_DEFAULT_STD_C(XX) less fragile, since clang-side refactors
caused problems there.

This uses ObjC and ObjCXX to provide the language standards, since that
matched the existing language standard behavior here.