[SwiftSyntax] Add an API to enable or disable bare slash regex literals from SwiftSyntax
This adds C functions that can be called from SwiftSyntax to enable or disable bare slash regex parsing.

It also adds a function to set the Swift language version - while this is not of top priority at the moment it will become important once we introduce a new language mode again.

rdar://93750821