More fixit options for unused async-let
This improves the fixits for unused async-let declarations.
When working with an async-let, the behavior of `await`ing the unused
variable vs not has an impact on the code in ways that may be
surprising.

The fixits give the option of renaming the variable `_`, which will
cancel the execution of the task and await it. Alternatively, awaiting
the task explicitly will allow it to run to completion without
cancellation.